backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
numpy==1.24.4
opencv-python-headless==4.8.1.78
Pillow==10.3.0
requests==2.32.3
onnxruntime==1.17.3
cachetools==5.3.3
flask-cors>=4.0.0